requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
arrayfire==3.8.0
ascii==3.6
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cached-property==2.0.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
cocos==0.2.2
colorama==0.4.6
comm==0.2.2
contexttimer==0.3.3
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.15
decorator==5.2.1
dill==0.4.0
distro==1.9.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
filelock==3.18.0
Flask==3.1.1
flet==0.28.3
flet-desktop==0.28.3
flet-web==0.28.3
fonttools==4.59.0
fsspec==2025.7.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
loky==3.5.5
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
oauthlib==3.3.1
openai==1.97.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
Panda3D==1.10.15
panda3d-gltf==1.3.0
panda3d-simplepbr==0.13.1
pandas==2.3.1
parso==0.8.4
pathos==0.3.4
pefile==2023.2.7
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pox==0.3.6
ppft==1.7.7
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyglet==2.1.6
Pygments==2.19.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.3
pyperclip==1.9.0
pyserial==3.5
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
repath==0.9.0
requests==2.32.4
safetensors==0.5.3
scikit-build==0.18.1
scipy==1.16.1
screeninfo==0.8.1
selenium==4.34.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sswg==4.3.0
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
ursina==8.0.1
uvicorn==0.35.0
validators==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
